Move all config values back to single file
Temporary intermediate step.
