+++ __tests__/applyAtRule.test.js
@@ -1,10 +1,9 @@
 import postcss from 'postcss'
 import substituteClassApplyAtRules from '../src/lib/substituteClassApplyAtRules'
 import processPlugins from '../src/util/processPlugins'
-import defaultPlugins from '../defaultPlugins'
-import defaultConfig from '../legacyConfig.stub.js'
+import defaultConfig from '../defaultConfig.stub.js'
 
-const { utilities: defaultUtilities } = processPlugins(defaultPlugins(), defaultConfig)
+const { utilities: defaultUtilities } = processPlugins(defaultConfig.plugins, defaultConfig)
 
 function run(input, config = defaultConfig, utilities = defaultUtilities) {
   return postcss([substituteClassApplyAtRules(config, utilities)]).process(input, {
@@ -205,10 +204,12 @@ test('you can apply utility classes without using the given prefix', () => {
     prefix: 'tw-',
   }
 
-  return run(input, config, processPlugins(defaultPlugins(), config).utilities).then(result => {
-    expect(result.css).toEqual(expected)
-    expect(result.warnings().length).toBe(0)
-  })
+  return run(input, config, processPlugins(defaultConfig.plugins, config).utilities).then(
+    result => {
+      expect(result.css).toEqual(expected)
+      expect(result.warnings().length).toBe(0)
+    }
+  )
 })
 
 test('you can apply utility classes without using the given prefix when using a function for the prefix', () => {
@@ -227,8 +228,10 @@ test('you can apply utility classes without using the given prefix when using a 
     },
   }
 
-  return run(input, config, processPlugins(defaultPlugins(), config).utilities).then(result => {
-    expect(result.css).toEqual(expected)
-    expect(result.warnings().length).toBe(0)
-  })
+  return run(input, config, processPlugins(defaultConfig.plugins, config).utilities).then(
+    result => {
+      expect(result.css).toEqual(expected)
+      expect(result.warnings().length).toBe(0)
+    }
+  )
 })
