+++ defaultConfig.stub.js
@@ -1,10 +1,835 @@
-const scales = require('./defaultScales')
+const colors = {
+  transparent: 'transparent',
+
+  'black': '#22292f',
+  'grey-darkest': '#3d4852',
+  'grey-darker': '#606f7b',
+  'grey-dark': '#8795a1',
+  'grey': '#b8c2cc',
+  'grey-light': '#dae1e7',
+  'grey-lighter': '#f1f5f8',
+  'grey-lightest': '#f8fafc',
+  'white': '#ffffff',
+
+  'red-darkest': '#3b0d0c',
+  'red-darker': '#621b18',
+  'red-dark': '#cc1f1a',
+  'red': '#e3342f',
+  'red-light': '#ef5753',
+  'red-lighter': '#f9acaa',
+  'red-lightest': '#fcebea',
+
+  'orange-darkest': '#462a16',
+  'orange-darker': '#613b1f',
+  'orange-dark': '#de751f',
+  'orange': '#f6993f',
+  'orange-light': '#faad63',
+  'orange-lighter': '#fcd9b6',
+  'orange-lightest': '#fff5eb',
+
+  'yellow-darkest': '#453411',
+  'yellow-darker': '#684f1d',
+  'yellow-dark': '#f2d024',
+  'yellow': '#ffed4a',
+  'yellow-light': '#fff382',
+  'yellow-lighter': '#fff9c2',
+  'yellow-lightest': '#fcfbeb',
+
+  'green-darkest': '#0f2f21',
+  'green-darker': '#1a4731',
+  'green-dark': '#1f9d55',
+  'green': '#38c172',
+  'green-light': '#51d88a',
+  'green-lighter': '#a2f5bf',
+  'green-lightest': '#e3fcec',
+
+  'teal-darkest': '#0d3331',
+  'teal-darker': '#20504f',
+  'teal-dark': '#38a89d',
+  'teal': '#4dc0b5',
+  'teal-light': '#64d5ca',
+  'teal-lighter': '#a0f0ed',
+  'teal-lightest': '#e8fffe',
+
+  'blue-darkest': '#12283a',
+  'blue-darker': '#1c3d5a',
+  'blue-dark': '#2779bd',
+  'blue': '#3490dc',
+  'blue-light': '#6cb2eb',
+  'blue-lighter': '#bcdefa',
+  'blue-lightest': '#eff8ff',
+
+  'indigo-darkest': '#191e38',
+  'indigo-darker': '#2f365f',
+  'indigo-dark': '#5661b3',
+  'indigo': '#6574cd',
+  'indigo-light': '#7886d7',
+  'indigo-lighter': '#b2b7ff',
+  'indigo-lightest': '#e6e8ff',
+
+  'purple-darkest': '#21183c',
+  'purple-darker': '#382b5f',
+  'purple-dark': '#794acf',
+  'purple': '#9561e2',
+  'purple-light': '#a779e9',
+  'purple-lighter': '#d6bbfc',
+  'purple-lightest': '#f3ebff',
+
+  'pink-darkest': '#451225',
+  'pink-darker': '#6f213f',
+  'pink-dark': '#eb5286',
+  'pink': '#f66d9b',
+  'pink-light': '#fa7ea8',
+  'pink-lighter': '#ffbbca',
+  'pink-lightest': '#ffebef',
+}
+
+const styles = {
+  /*
+  |-----------------------------------------------------------------------------
+  | Fonts                                    https://tailwindcss.com/docs/fonts
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your project's font stack, or font families.
+  | Keep in mind that Tailwind doesn't actually load any fonts for you.
+  | If you're using custom fonts you'll need to import them prior to
+  | defining them here.
+  |
+  | By default we provide a native font stack that works remarkably well on
+  | any device or OS you're using, since it just uses the default fonts
+  | provided by the platform.
+  |
+  | Class name: .font-{name}
+  | CSS property: font-family
+  |
+  */
+
+  fonts: {
+    sans: [
+      'system-ui',
+      'BlinkMacSystemFont',
+      '-apple-system',
+      'Segoe UI',
+      'Roboto',
+      'Oxygen',
+      'Ubuntu',
+      'Cantarell',
+      'Fira Sans',
+      'Droid Sans',
+      'Helvetica Neue',
+      'sans-serif',
+    ],
+    serif: [
+      'Constantia',
+      'Lucida Bright',
+      'Lucidabright',
+      'Lucida Serif',
+      'Lucida',
+      'DejaVu Serif',
+      'Bitstream Vera Serif',
+      'Liberation Serif',
+      'Georgia',
+      'serif',
+    ],
+    mono: [
+      'Menlo',
+      'Monaco',
+      'Consolas',
+      'Liberation Mono',
+      'Courier New',
+      'monospace'
+    ],
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Text sizes                         https://tailwindcss.com/docs/text-sizing
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your text sizes. Name these in whatever way
+  | makes the most sense to you. We use size names by default, but
+  | you're welcome to use a numeric scale or even something else
+  | entirely.
+  |
+  | By default Tailwind uses the "rem" unit type for most measurements.
+  | This allows you to set a root font size which all other sizes are
+  | then based on. That said, you are free to use whatever units you
+  | prefer, be it rems, ems, pixels or other.
+  |
+  | Class name: .text-{size}
+  | CSS property: font-size
+  |
+  */
+
+  textSizes: {
+    xs: '.75rem', // 12px
+    sm: '.875rem', // 14px
+    base: '1rem', // 16px
+    lg: '1.125rem', // 18px
+    xl: '1.25rem', // 20px
+    '2xl': '1.5rem', // 24px
+    '3xl': '1.875rem', // 30px
+    '4xl': '2.25rem', // 36px
+    '5xl': '3rem', // 48px
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Font weights                       https://tailwindcss.com/docs/font-weight
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your font weights. We've provided a list of
+  | common font weight names with their respective numeric scale values
+  | to get you started. It's unlikely that your project will require
+  | all of these, so we recommend removing those you don't need.
+  |
+  | Class name: .font-{weight}
+  | CSS property: font-weight
+  |
+  */
+
+  fontWeights: {
+    hairline: 100,
+    thin: 200,
+    light: 300,
+    normal: 400,
+    medium: 500,
+    semibold: 600,
+    bold: 700,
+    extrabold: 800,
+    black: 900,
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Leading (line height)              https://tailwindcss.com/docs/line-height
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your line height values, or as we call
+  | them in Tailwind, leadings.
+  |
+  | Class name: .leading-{size}
+  | CSS property: line-height
+  |
+  */
+
+  leading: {
+    none: 1,
+    tight: 1.25,
+    normal: 1.5,
+    loose: 2,
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Tracking (letter spacing)       https://tailwindcss.com/docs/letter-spacing
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your letter spacing values, or as we call
+  | them in Tailwind, tracking.
+  |
+  | Class name: .tracking-{size}
+  | CSS property: letter-spacing
+  |
+  */
+
+  tracking: {
+    tight: '-0.05em',
+    normal: '0',
+    wide: '0.05em',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Text colors                         https://tailwindcss.com/docs/text-color
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your text colors. By default these use the
+  | color palette we defined above, however you're welcome to set these
+  | independently if that makes sense for your project.
+  |
+  | Class name: .text-{color}
+  | CSS property: color
+  |
+  */
+
+  textColors: colors,
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Background colors             https://tailwindcss.com/docs/background-color
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your background colors. By default these use
+  | the color palette we defined above, however you're welcome to set
+  | these independently if that makes sense for your project.
+  |
+  | Class name: .bg-{color}
+  | CSS property: background-color
+  |
+  */
+
+  backgroundColors: colors,
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Background sizes               https://tailwindcss.com/docs/background-size
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your background sizes. We provide some common
+  | values that are useful in most projects, but feel free to add other sizes
+  | that are specific to your project here as well.
+  |
+  | Class name: .bg-{size}
+  | CSS property: background-size
+  |
+  */
+
+  backgroundSize: {
+    auto: 'auto',
+    cover: 'cover',
+    contain: 'contain',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Border widths                     https://tailwindcss.com/docs/border-width
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your border widths. Take note that border
+  | widths require a special "default" value set as well. This is the
+  | width that will be used when you do not specify a border width.
+  |
+  | Class name: .border{-side?}{-width?}
+  | CSS property: border-width
+  |
+  */
+
+  borderWidths: {
+    default: '1px',
+    '0': '0',
+    '2': '2px',
+    '4': '4px',
+    '8': '8px',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Border colors                     https://tailwindcss.com/docs/border-color
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your border colors. By default these use the
+  | color palette we defined above, however you're welcome to set these
+  | independently if that makes sense for your project.
+  |
+  | Take note that border colors require a special "default" value set
+  | as well. This is the color that will be used when you do not
+  | specify a border color.
+  |
+  | Class name: .border-{color}
+  | CSS property: border-color
+  |
+  */
+
+  borderColors: global.Object.assign({ default: colors['grey-light'] }, colors),
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Border radius                    https://tailwindcss.com/docs/border-radius
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your border radius values. If a `default` radius
+  | is provided, it will be made available as the non-suffixed `.rounded`
+  | utility.
+  |
+  | If your scale includes a `0` value to reset already rounded corners, it's
+  | a good idea to put it first so other values are able to override it.
+  |
+  | Class name: .rounded{-side?}{-size?}
+  | CSS property: border-radius
+  |
+  */
+
+  borderRadius: {
+    none: '0',
+    sm: '.125rem',
+    default: '.25rem',
+    lg: '.5rem',
+    full: '9999px',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Width                                    https://tailwindcss.com/docs/width
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your width utility sizes. These can be
+  | percentage based, pixels, rems, or any other units. By default
+  | we provide a sensible rem based numeric scale, a percentage
+  | based fraction scale, plus some other common use-cases. You
+  | can, of course, modify these values as needed.
+  |
+  |
+  | It's also worth mentioning that Tailwind automatically escapes
+  | invalid CSS class name characters, which allows you to have
+  | awesome classes like .w-2/3.
+  |
+  | Class name: .w-{size}
+  | CSS property: width
+  |
+  */
+
+  width: {
+    auto: 'auto',
+    px: '1px',
+    '1': '0.25rem',
+    '2': '0.5rem',
+    '3': '0.75rem',
+    '4': '1rem',
+    '5': '1.25rem',
+    '6': '1.5rem',
+    '8': '2rem',
+    '10': '2.5rem',
+    '12': '3rem',
+    '16': '4rem',
+    '24': '6rem',
+    '32': '8rem',
+    '48': '12rem',
+    '64': '16rem',
+    '1/2': '50%',
+    '1/3': '33.33333%',
+    '2/3': '66.66667%',
+    '1/4': '25%',
+    '3/4': '75%',
+    '1/5': '20%',
+    '2/5': '40%',
+    '3/5': '60%',
+    '4/5': '80%',
+    '1/6': '16.66667%',
+    '5/6': '83.33333%',
+    full: '100%',
+    screen: '100vw',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Height                                  https://tailwindcss.com/docs/height
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your height utility sizes. These can be
+  | percentage based, pixels, rems, or any other units. By default
+  | we provide a sensible rem based numeric scale plus some other
+  | common use-cases. You can, of course, modify these values as
+  | needed.
+  |
+  | Class name: .h-{size}
+  | CSS property: height
+  |
+  */
+
+  height: {
+    auto: 'auto',
+    px: '1px',
+    '1': '0.25rem',
+    '2': '0.5rem',
+    '3': '0.75rem',
+    '4': '1rem',
+    '5': '1.25rem',
+    '6': '1.5rem',
+    '8': '2rem',
+    '10': '2.5rem',
+    '12': '3rem',
+    '16': '4rem',
+    '24': '6rem',
+    '32': '8rem',
+    '48': '12rem',
+    '64': '16rem',
+    full: '100%',
+    screen: '100vh',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Minimum width                        https://tailwindcss.com/docs/min-width
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your minimum width utility sizes. These can
+  | be percentage based, pixels, rems, or any other units. We provide a
+  | couple common use-cases by default. You can, of course, modify
+  | these values as needed.
+  |
+  | Class name: .min-w-{size}
+  | CSS property: min-width
+  |
+  */
+
+  minWidth: {
+    '0': '0',
+    full: '100%',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Minimum height                      https://tailwindcss.com/docs/min-height
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your minimum height utility sizes. These can
+  | be percentage based, pixels, rems, or any other units. We provide a
+  | few common use-cases by default. You can, of course, modify these
+  | values as needed.
+  |
+  | Class name: .min-h-{size}
+  | CSS property: min-height
+  |
+  */
+
+  minHeight: {
+    '0': '0',
+    full: '100%',
+    screen: '100vh',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Maximum width                        https://tailwindcss.com/docs/max-width
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your maximum width utility sizes. These can
+  | be percentage based, pixels, rems, or any other units. By default
+  | we provide a sensible rem based scale and a "full width" size,
+  | which is basically a reset utility. You can, of course,
+  | modify these values as needed.
+  |
+  | Class name: .max-w-{size}
+  | CSS property: max-width
+  |
+  */
+
+  maxWidth: {
+    xs: '20rem',
+    sm: '30rem',
+    md: '40rem',
+    lg: '50rem',
+    xl: '60rem',
+    '2xl': '70rem',
+    '3xl': '80rem',
+    '4xl': '90rem',
+    '5xl': '100rem',
+    full: '100%',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Maximum height                      https://tailwindcss.com/docs/max-height
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your maximum height utility sizes. These can
+  | be percentage based, pixels, rems, or any other units. We provide a
+  | couple common use-cases by default. You can, of course, modify
+  | these values as needed.
+  |
+  | Class name: .max-h-{size}
+  | CSS property: max-height
+  |
+  */
+
+  maxHeight: {
+    full: '100%',
+    screen: '100vh',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Padding                                https://tailwindcss.com/docs/padding
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your padding utility sizes. These can be
+  | percentage based, pixels, rems, or any other units. By default we
+  | provide a sensible rem based numeric scale plus a couple other
+  | common use-cases like "1px". You can, of course, modify these
+  | values as needed.
+  |
+  | Class name: .p{side?}-{size}
+  | CSS property: padding
+  |
+  */
+
+  padding: {
+    px: '1px',
+    '0': '0',
+    '1': '0.25rem',
+    '2': '0.5rem',
+    '3': '0.75rem',
+    '4': '1rem',
+    '5': '1.25rem',
+    '6': '1.5rem',
+    '8': '2rem',
+    '10': '2.5rem',
+    '12': '3rem',
+    '16': '4rem',
+    '20': '5rem',
+    '24': '6rem',
+    '32': '8rem',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Margin                                  https://tailwindcss.com/docs/margin
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your margin utility sizes. These can be
+  | percentage based, pixels, rems, or any other units. By default we
+  | provide a sensible rem based numeric scale plus a couple other
+  | common use-cases like "1px". You can, of course, modify these
+  | values as needed.
+  |
+  | Class name: .m{side?}-{size}
+  | CSS property: margin
+  |
+  */
+
+  margin: {
+    auto: 'auto',
+    px: '1px',
+    '0': '0',
+    '1': '0.25rem',
+    '2': '0.5rem',
+    '3': '0.75rem',
+    '4': '1rem',
+    '5': '1.25rem',
+    '6': '1.5rem',
+    '8': '2rem',
+    '10': '2.5rem',
+    '12': '3rem',
+    '16': '4rem',
+    '20': '5rem',
+    '24': '6rem',
+    '32': '8rem',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Negative margin                https://tailwindcss.com/docs/negative-margin
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your negative margin utility sizes. These can
+  | be percentage based, pixels, rems, or any other units. By default we
+  | provide matching values to the padding scale since these utilities
+  | generally get used together. You can, of course, modify these
+  | values as needed.
+  |
+  | Class name: .-m{side?}-{size}
+  | CSS property: margin
+  |
+  */
+
+  negativeMargin: {
+    px: '1px',
+    '0': '0',
+    '1': '0.25rem',
+    '2': '0.5rem',
+    '3': '0.75rem',
+    '4': '1rem',
+    '5': '1.25rem',
+    '6': '1.5rem',
+    '8': '2rem',
+    '10': '2.5rem',
+    '12': '3rem',
+    '16': '4rem',
+    '20': '5rem',
+    '24': '6rem',
+    '32': '8rem',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Shadows                                https://tailwindcss.com/docs/shadows
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your shadow utilities. As you can see from
+  | the defaults we provide, it's possible to apply multiple shadows
+  | per utility using comma separation.
+  |
+  | If a `default` shadow is provided, it will be made available as the non-
+  | suffixed `.shadow` utility.
+  |
+  | Class name: .shadow-{size?}
+  | CSS property: box-shadow
+  |
+  */
+
+  shadows: {
+    default: '0 2px 4px 0 rgba(0,0,0,0.10)',
+    md: '0 4px 8px 0 rgba(0,0,0,0.12), 0 2px 4px 0 rgba(0,0,0,0.08)',
+    lg: '0 15px 30px 0 rgba(0,0,0,0.11), 0 5px 15px 0 rgba(0,0,0,0.08)',
+    inner: 'inset 0 2px 4px 0 rgba(0,0,0,0.06)',
+    outline: '0 0 0 3px rgba(52,144,220,0.5)',
+    none: 'none',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Z-index                                https://tailwindcss.com/docs/z-index
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your z-index utility values. By default we
+  | provide a sensible numeric scale. You can, of course, modify these
+  | values as needed.
+  |
+  | Class name: .z-{index}
+  | CSS property: z-index
+  |
+  */
+
+  zIndex: {
+    auto: 'auto',
+    '0': 0,
+    '10': 10,
+    '20': 20,
+    '30': 30,
+    '40': 40,
+    '50': 50,
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Opacity                                https://tailwindcss.com/docs/opacity
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your opacity utility values. By default we
+  | provide a sensible numeric scale. You can, of course, modify these
+  | values as needed.
+  |
+  | Class name: .opacity-{name}
+  | CSS property: opacity
+  |
+  */
+
+  opacity: {
+    '0': '0',
+    '25': '.25',
+    '50': '.5',
+    '75': '.75',
+    '100': '1',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | SVG fill                                   https://tailwindcss.com/docs/svg
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your SVG fill colors. By default we just provide
+  | `fill-current` which sets the fill to the current text color. This lets you
+  | specify a fill color using existing text color utilities and helps keep the
+  | generated CSS file size down.
+  |
+  | Class name: .fill-{name}
+  | CSS property: fill
+  |
+  */
+
+  svgFill: {
+    current: 'currentColor',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | SVG stroke                                 https://tailwindcss.com/docs/svg
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you define your SVG stroke colors. By default we just provide
+  | `stroke-current` which sets the stroke to the current text color. This lets
+  | you specify a stroke color using existing text color utilities and helps
+  | keep the generated CSS file size down.
+  |
+  | Class name: .stroke-{name}
+  | CSS property: stroke
+  |
+  */
+
+  svgStroke: {
+    current: 'currentColor',
+  },
+
+  /*
+  |-----------------------------------------------------------------------------
+  | Modules                  https://tailwindcss.com/docs/configuration#modules
+  |-----------------------------------------------------------------------------
+  |
+  | Here is where you control which modules are generated and what variants are
+  | generated for each of those modules.
+  |
+  | Currently supported variants:
+  |   - responsive
+  |   - hover
+  |   - focus
+  |   - focus-within
+  |   - active
+  |   - group-hover
+  |
+  | To disable a module completely, use `false` instead of an array.
+  |
+  */
+
+  modules: {
+    appearance: ['responsive'],
+    backgroundAttachment: ['responsive'],
+    backgroundColors: ['responsive', 'hover', 'focus'],
+    backgroundPosition: ['responsive'],
+    backgroundRepeat: ['responsive'],
+    backgroundSize: ['responsive'],
+    borderCollapse: [],
+    borderColors: ['responsive', 'hover', 'focus'],
+    borderRadius: ['responsive'],
+    borderStyle: ['responsive'],
+    borderWidths: ['responsive'],
+    cursor: ['responsive'],
+    display: ['responsive'],
+    flexbox: ['responsive'],
+    float: ['responsive'],
+    fonts: ['responsive'],
+    fontWeights: ['responsive', 'hover', 'focus'],
+    height: ['responsive'],
+    leading: ['responsive'],
+    lists: ['responsive'],
+    margin: ['responsive'],
+    maxHeight: ['responsive'],
+    maxWidth: ['responsive'],
+    minHeight: ['responsive'],
+    minWidth: ['responsive'],
+    negativeMargin: ['responsive'],
+    objectFit: false,
+    objectPosition: false,
+    opacity: ['responsive'],
+    outline: ['focus'],
+    overflow: ['responsive'],
+    padding: ['responsive'],
+    pointerEvents: ['responsive'],
+    position: ['responsive'],
+    resize: ['responsive'],
+    shadows: ['responsive', 'hover', 'focus'],
+    svgFill: [],
+    svgStroke: [],
+    tableLayout: ['responsive'],
+    textAlign: ['responsive'],
+    textColors: ['responsive', 'hover', 'focus'],
+    textSizes: ['responsive'],
+    textStyle: ['responsive', 'hover', 'focus'],
+    tracking: ['responsive'],
+    userSelect: ['responsive'],
+    verticalAlign: ['responsive'],
+    visibility: ['responsive'],
+    whitespace: ['responsive'],
+    width: ['responsive'],
+    zIndex: ['responsive'],
+  },
+}
 
 module.exports = {
   prefix: '',
   important: false,
   separator: ':',
-  colors: scales.colors,
+  colors: colors,
   screens: {
     'sm': '576px',
     'md': '768px',
@@ -16,55 +841,55 @@ module.exports = {
       // center: true,
       // padding: '1rem',
     }),
-    require('./plugins/lists')({ variants: scales.modules.lists }),
-    require('./plugins/appearance')({ variants: scales.modules.appearance }),
-    require('./plugins/backgroundAttachment')({ variants: scales.modules.backgroundAttachment }),
-    require('./plugins/backgroundColors')({ values: scales.backgroundColors, variants: scales.modules.backgroundColors }),
-    require('./plugins/backgroundPosition')({ variants: scales.modules.backgroundPosition }),
-    require('./plugins/backgroundRepeat')({ variants: scales.modules.backgroundRepeat }),
-    require('./plugins/backgroundSize')({ values: scales.backgroundSize, variants: scales.modules.backgroundSize }),
-    require('./plugins/borderCollapse')({ variants: scales.modules.borderCollapse }),
-    require('./plugins/borderColors')({ values: scales.borderColors, variants: scales.modules.borderColors }),
-    require('./plugins/borderRadius')({ values: scales.borderRadius, variants: scales.modules.borderRadius }),
-    require('./plugins/borderStyle')({ variants: scales.modules.borderStyle }),
-    require('./plugins/borderWidths')({ values: scales.borderWidths, variants: scales.modules.borderWidths }),
-    require('./plugins/cursor')({ variants: scales.modules.cursor }),
-    require('./plugins/display')({ variants: scales.modules.display }),
-    require('./plugins/flexbox')({ variants: scales.modules.flexbox }),
-    require('./plugins/float')({ variants: scales.modules.float }),
-    require('./plugins/fonts')({ values: scales.fonts, variants: scales.modules.fonts }),
-    require('./plugins/fontWeights')({ values: scales.fontWeights, variants: scales.modules.fontWeights }),
-    require('./plugins/height')({ values: scales.height, variants: scales.modules.height }),
-    require('./plugins/leading')({ values: scales.leading, variants: scales.modules.leading }),
-    require('./plugins/margin')({ values: scales.margin, variants: scales.modules.margin }),
-    require('./plugins/maxHeight')({ values: scales.maxHeight, variants: scales.modules.maxHeight }),
-    require('./plugins/maxWidth')({ values: scales.maxWidth, variants: scales.modules.maxWidth }),
-    require('./plugins/minHeight')({ values: scales.minHeight, variants: scales.modules.minHeight }),
-    require('./plugins/minWidth')({ values: scales.minWidth, variants: scales.modules.minWidth }),
-    require('./plugins/negativeMargin')({ values: scales.negativeMargin, variants: scales.modules.negativeMargin }),
-    // require('./plugins/objectFit')({ variants: scales.modules.objectFit }),
-    // require('./plugins/objectPosition')({ variants: scales.modules.objectPosition }),
-    require('./plugins/opacity')({ values: scales.opacity, variants: scales.modules.opacity }),
-    require('./plugins/outline')({ variants: scales.modules.outline }),
-    require('./plugins/overflow')({ variants: scales.modules.overflow }),
-    require('./plugins/padding')({ values: scales.padding, variants: scales.modules.padding }),
-    require('./plugins/pointerEvents')({ variants: scales.modules.pointerEvents }),
-    require('./plugins/position')({ variants: scales.modules.position }),
-    require('./plugins/resize')({ variants: scales.modules.resize }),
-    require('./plugins/shadows')({ values: scales.shadows, variants: scales.modules.shadows }),
-    require('./plugins/svgFill')({ values: scales.svgFill, variants: scales.modules.svgFill }),
-    require('./plugins/svgStroke')({ values: scales.svgStroke, variants: scales.modules.svgStroke }),
-    require('./plugins/tableLayout')({ variants: scales.modules.tableLayout }),
-    require('./plugins/textAlign')({ variants: scales.modules.textAlign }),
-    require('./plugins/textColors')({ values: scales.textColors, variants: scales.modules.textColors }),
-    require('./plugins/textSizes')({ values: scales.textSizes, variants: scales.modules.textSizes }),
-    require('./plugins/textStyle')({ variants: scales.modules.textStyle }),
-    require('./plugins/tracking')({ values: scales.tracking, variants: scales.modules.tracking }),
-    require('./plugins/userSelect')({ variants: scales.modules.userSelect }),
-    require('./plugins/verticalAlign')({ variants: scales.modules.verticalAlign }),
-    require('./plugins/visibility')({ variants: scales.modules.visibility }),
-    require('./plugins/whitespace')({ variants: scales.modules.whitespace }),
-    require('./plugins/width')({ values: scales.width, variants: scales.modules.width }),
-    require('./plugins/zIndex')({ values: scales.zIndex, variants: scales.modules.zIndex }),
+    require('./plugins/lists')({ variants: styles.modules.lists }),
+    require('./plugins/appearance')({ variants: styles.modules.appearance }),
+    require('./plugins/backgroundAttachment')({ variants: styles.modules.backgroundAttachment }),
+    require('./plugins/backgroundColors')({ values: styles.backgroundColors, variants: styles.modules.backgroundColors }),
+    require('./plugins/backgroundPosition')({ variants: styles.modules.backgroundPosition }),
+    require('./plugins/backgroundRepeat')({ variants: styles.modules.backgroundRepeat }),
+    require('./plugins/backgroundSize')({ values: styles.backgroundSize, variants: styles.modules.backgroundSize }),
+    require('./plugins/borderCollapse')({ variants: styles.modules.borderCollapse }),
+    require('./plugins/borderColors')({ values: styles.borderColors, variants: styles.modules.borderColors }),
+    require('./plugins/borderRadius')({ values: styles.borderRadius, variants: styles.modules.borderRadius }),
+    require('./plugins/borderStyle')({ variants: styles.modules.borderStyle }),
+    require('./plugins/borderWidths')({ values: styles.borderWidths, variants: styles.modules.borderWidths }),
+    require('./plugins/cursor')({ variants: styles.modules.cursor }),
+    require('./plugins/display')({ variants: styles.modules.display }),
+    require('./plugins/flexbox')({ variants: styles.modules.flexbox }),
+    require('./plugins/float')({ variants: styles.modules.float }),
+    require('./plugins/fonts')({ values: styles.fonts, variants: styles.modules.fonts }),
+    require('./plugins/fontWeights')({ values: styles.fontWeights, variants: styles.modules.fontWeights }),
+    require('./plugins/height')({ values: styles.height, variants: styles.modules.height }),
+    require('./plugins/leading')({ values: styles.leading, variants: styles.modules.leading }),
+    require('./plugins/margin')({ values: styles.margin, variants: styles.modules.margin }),
+    require('./plugins/maxHeight')({ values: styles.maxHeight, variants: styles.modules.maxHeight }),
+    require('./plugins/maxWidth')({ values: styles.maxWidth, variants: styles.modules.maxWidth }),
+    require('./plugins/minHeight')({ values: styles.minHeight, variants: styles.modules.minHeight }),
+    require('./plugins/minWidth')({ values: styles.minWidth, variants: styles.modules.minWidth }),
+    require('./plugins/negativeMargin')({ values: styles.negativeMargin, variants: styles.modules.negativeMargin }),
+    // require('./plugins/objectFit')({ variants: styles.modules.objectFit }),
+    // require('./plugins/objectPosition')({ variants: styles.modules.objectPosition }),
+    require('./plugins/opacity')({ values: styles.opacity, variants: styles.modules.opacity }),
+    require('./plugins/outline')({ variants: styles.modules.outline }),
+    require('./plugins/overflow')({ variants: styles.modules.overflow }),
+    require('./plugins/padding')({ values: styles.padding, variants: styles.modules.padding }),
+    require('./plugins/pointerEvents')({ variants: styles.modules.pointerEvents }),
+    require('./plugins/position')({ variants: styles.modules.position }),
+    require('./plugins/resize')({ variants: styles.modules.resize }),
+    require('./plugins/shadows')({ values: styles.shadows, variants: styles.modules.shadows }),
+    require('./plugins/svgFill')({ values: styles.svgFill, variants: styles.modules.svgFill }),
+    require('./plugins/svgStroke')({ values: styles.svgStroke, variants: styles.modules.svgStroke }),
+    require('./plugins/tableLayout')({ variants: styles.modules.tableLayout }),
+    require('./plugins/textAlign')({ variants: styles.modules.textAlign }),
+    require('./plugins/textColors')({ values: styles.textColors, variants: styles.modules.textColors }),
+    require('./plugins/textSizes')({ values: styles.textSizes, variants: styles.modules.textSizes }),
+    require('./plugins/textStyle')({ variants: styles.modules.textStyle }),
+    require('./plugins/tracking')({ values: styles.tracking, variants: styles.modules.tracking }),
+    require('./plugins/userSelect')({ variants: styles.modules.userSelect }),
+    require('./plugins/verticalAlign')({ variants: styles.modules.verticalAlign }),
+    require('./plugins/visibility')({ variants: styles.modules.visibility }),
+    require('./plugins/whitespace')({ variants: styles.modules.whitespace }),
+    require('./plugins/width')({ values: styles.width, variants: styles.modules.width }),
+    require('./plugins/zIndex')({ values: styles.zIndex, variants: styles.modules.zIndex }),
   ],
 }
